+++ activesupport/lib/active_support.rb
@@ -79,14 +79,6 @@ module ActiveSupport
 
   cattr_accessor :test_order # :nodoc:
 
-  def self.halt_callback_chains_on_return_false
-    Callbacks.halt_and_display_warning_on_return_false
-  end
-
-  def self.halt_callback_chains_on_return_false=(value)
-    Callbacks.halt_and_display_warning_on_return_false = value
-  end
-
   def self.to_time_preserves_timezone
     DateAndTime::Compatibility.preserve_timezone
   end