+++ activemodel/CHANGELOG.md
@@ -1,2 +1,5 @@
+*   Removed support for returning false to halt validation callbacks (deprecated).
+
+    *Iain Beeston*
 
 Please check [5-0-stable](https://github.com/rails/rails/blob/5-0-stable/activemodel/CHANGELOG.md) for previous changes.